+++ src/sgg_utils/sgg_utils.py
@@ -294,18 +294,6 @@ def get_bookings(token, course_id, teesheet_id, start_date, end_date = None, lim
 
     return bookings_data
 
-def access_secret_version(project_id, secret_id, version_id):
-    from google.cloud import secretmanager
-    # Create the Secret Manager client.
-    client = secretmanager.SecretManagerServiceClient()
 
-    # Build the resource name of the secret version.
-    name = f"projects/{project_id}/secrets/{secret_id}/versions/{version_id}"
-
-    # Access the secret version.
-    response = client.access_secret_version(request={"name": name})
-
-    # Return the decoded payload.
-    return response.payload.data.decode('UTF-8')
